# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Songguo Admin API
  version: "1"
  description: >
    Control-plane (admin) API for the Songguo AI usage gateway. It is read-mostly
    and consumed by the dashboard, the `songguo` MCP server, and any other tooling.
    Every endpoint is gated by a single admin bearer key (`SONGGUO_ADMIN_KEY`);
    when that key is empty the server runs the API unprotected and logs a warning.

    This describes only the control plane. The data plane — the transparent proxy
    at the native vendor paths (`/v1/*`, `/api/v3/*`), with the provider selected
    by the `X-Songguo-Provider` header, else the body's model, else the wire
    default — speaks each vendor's native wire protocol and is documented in
    docs/registry.md.
servers:
  - url: /
security:
  - adminKey: []
...
paths:
...
  /api/providers:
    get:
      tags: [providers]
      summary: List configured providers (keys masked)
      operationId: listProviders
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Provider" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [providers]
      summary: Create a provider
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProvider" }
      responses:
        "201":
          description: The created provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/providers/{id}:
    get:
      tags: [providers]
      summary: Get one provider
      operationId: getProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [providers]
      summary: Update a provider's mutable fields
      operationId: patchProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchProvider" }
      responses:
        "200":
          description: The updated provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [providers]
      summary: Delete a provider
      operationId: deleteProvider
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    adminKey:
      type: http
      scheme: bearer
      description: The admin key from SONGGUO_ADMIN_KEY.
...
  responses:
    Unauthorized:
      description: Missing or invalid admin key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict (e.g. duplicate name)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string, description: "songguo_-prefixed reason, e.g. songguo_not_found" }
...
    ProviderEndpoint:
      type: object
      properties:
        wire: { type: string, description: "Wire name, e.g. openai/chat." }
        endpoint: { type: string, description: "Full upstream URL for this wire." }
        adapter: { type: string, description: "Auth scheme: openai-compatible | anthropic-compatible | volc-speech." }
    ProviderModel:
      type: object
      properties:
        model: { type: string }
        input: { type: number }
        output: { type: number }
        cached_input: { type: number }
        unit: { type: string, description: "e.g. per_1m_tokens." }
        price_override: { type: boolean, description: "When false, catalog-backed providers use the current catalog price for this model." }
    VendorStats:
      type: object
      properties:
        requests: { type: integer }
        errors: { type: integer }
        error_rate: { type: number }
        avg_latency_ms: { type: number }
        last_status: { type: integer }
        healthy: { type: boolean }
    Provider:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        vendor: { type: string }
        priority: { type: integer }
        weight: { type: integer }
        enabled: { type: boolean }
        catalog_id: { type: string }
        endpoints: { type: array, items: { $ref: "#/components/schemas/ProviderEndpoint" } }
        allow_unmatched: { type: boolean }
        max_concurrency: { type: integer, description: "Max in-flight requests to this provider's credential; 0 = unlimited. Callers WAIT for a slot rather than being routed to another provider, which would discard the session prompt cache." }
        quirks: { type: object, additionalProperties: { type: string } }
        masked_key: { type: string }
        models: { type: array, items: { $ref: "#/components/schemas/ProviderModel" } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        stats: { $ref: "#/components/schemas/VendorStats" }
    CreateProvider:
      type: object
      required: [name]
      properties:
        name: { type: string }
        vendor: { type: string }
        priority: { type: integer }
        weight: { type: integer }
        enabled: { type: boolean }
        catalog_id: { type: string }
        allow_unmatched: { type: boolean }
        max_concurrency: { type: integer, description: "Max in-flight requests to this provider's credential; 0 = unlimited. Callers WAIT for a slot rather than being routed to another provider, which would discard the session prompt cache." }
        quirks: { type: object, additionalProperties: { type: string } }
        api_key: { type: string }
        models: { type: array, items: { $ref: "#/components/schemas/ProviderModel" } }
        endpoints: { type: array, items: { $ref: "#/components/schemas/ProviderEndpoint" } }
    PatchProvider:
      type: object
      description: Only present fields change. Supplying models or endpoints replaces those lists wholesale.
      properties:
        name: { type: string }
        vendor: { type: string }
        priority: { type: integer }
        weight: { type: integer }
        enabled: { type: boolean }
        allow_unmatched: { type: boolean }
        max_concurrency: { type: integer, description: "Max in-flight requests to this provider's credential; 0 = unlimited. Callers WAIT for a slot rather than being routed to another provider, which would discard the session prompt cache." }
        api_key: { type: string, description: "Replaces the key when non-empty." }
        quirks: { type: object, additionalProperties: { type: string } }
        models: { type: array, items: { $ref: "#/components/schemas/ProviderModel" } }
        endpoints: { type: array, items: { $ref: "#/components/schemas/ProviderEndpoint" } }
...